+++ template/src/frontend/index.tsx
@@ -5,18 +5,18 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { Logger, LogLevel } from "@bentley/bentleyjs-core";
-import { BasicViewportApp } from "./api/BasicViewportApp";
+import { iModeljsApp } from "./api/iModeljs";
 import App from "./components/App";
 import "./index.css";
 
-// Setup logging immediately to pick up any logging during BasicViewportApp.startup()
+// Setup logging immediately to pick up any logging during iModeljsApp.startup()
 Logger.initializeToConsole();
 Logger.setLevelDefault(LogLevel.Warning); // Set all logging to a default of Warning
-Logger.setLevel("basic-viewport-app", LogLevel.Info); // Override the above default and set only App level logging to Info.
+Logger.setLevel("iModeljs-app", LogLevel.Info); // Override the above default and set only App level logging to Info.
 
 (async () => {  // eslint-disable-line @typescript-eslint/no-floating-promises
   // initialize the application
-  await BasicViewportApp.startup();
+  await iModeljsApp.startup();
 
   // when initialization is complete, render
   ReactDOM.render(